# swagger.yaml
openapi: 3.0.0
info:
  title: Legal Advice
  contact: {}
  version: '1.0'
servers:
- url: https://legal-advice-1812.onrender.com/api/v1
  variables: {}
- url: http://localhost:1812/api/v1
  variables: {}
paths:
...
  /user/forgetPassword:
    post:
      tags:
      - User
      summary: Forget Password
      operationId: ForgetPassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/ForgetPasswordRequest'
              - example:
                  email: mh3908334@gmail.com
            example:
              email: mh3908334@gmail.com
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /user/verifyResetCode:
    post:
      tags:
      - User
      summary: Verify Reset Code
      operationId: VerifyResetCode
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/VerifyResetCodeRequest'
              - example:
                  email: mh3908334@gmail.com
                  resetCode: '573257'
            example:
              email: mh3908334@gmail.com
              resetCode: '573257'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /user/resetPassword:
    post:
      tags:
      - User
      summary: Reset Password
      operationId: ResetPassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/ResetPasswordRequest'
              - example:
                  newPassword: '123456'
                  email: mh3908334@gmail.com
            example:
              newPassword: '123456'
              email: mh3908334@gmail.com
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /user/{id}:
    get:
      tags:
      - About Me
      summary: Get About Me
      operationId: GetAboutMe
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    patch:
      tags:
      - About Me
      summary: Update About Me
      operationId: UpdateAboutMe
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              - name
              - email
              - phone
              - aboutMe
              - specializations
              - experiences
              type: object
              properties:
                image:
                  type: string
                  format: binary
                name:
                  type: string
                  example: Test
                email:
                  type: string
                  example: test@test.com
                phone:
                  type: string
                  example: '01154604708'
                aboutMe:
                  type: string
                  example: Test
                specializations:
                  type: string
                  example: Test
                experiences:
                  type: string
                  example: Test
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /user/admin/{id}:
    post:
      tags:
      - About Me
      summary: About Me
      operationId: AboutMe
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              - name
              - email
              - phone
              - aboutMe
              - specializations
              - qualifications
              - experiences
              type: object
              properties:
                image:
                  type: string
                  format: binary
                name:
                  type: string
                  example: Admin
                email:
                  type: string
                  example: mh3908334@test.com
                phone:
                  type: string
                  example: '01154604708'
                aboutMe:
                  type: string
                  example: Test
                specializations:
                  type: string
                  example: Test 1
                qualifications:
                  type: string
                  example: Test 1
                experiences:
                  type: string
                  example: Test 1
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /user/admin/signup:
    patch:
      tags:
      - About Me
      summary: Signup As Admin
      operationId: SignupAsAdmin
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/SignupAsAdminRequest'
              - example:
                  name: Admin
                  email: admin@gmail.com
                  password: '123456'
            example:
              name: Admin
              email: admin@gmail.com
              password: '123456'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    ForgetPasswordRequest:
      title: ForgetPasswordRequest
      required:
      - email
      type: object
      properties:
        email:
          type: string
      example:
        email: mh3908334@gmail.com
    VerifyResetCodeRequest:
      title: VerifyResetCodeRequest
      required:
      - email
      - resetCode
      type: object
      properties:
        email:
          type: string
        resetCode:
          type: string
      example:
        email: mh3908334@gmail.com
        resetCode: '573257'
    ResetPasswordRequest:
      title: ResetPasswordRequest
      required:
      - newPassword
      - email
      type: object
      properties:
        newPassword:
          type: string
        email:
          type: string
      example:
        newPassword: '123456'
        email: mh3908334@gmail.com
...
    SignupAsAdminRequest:
      title: SignupAsAdminRequest
      required:
      - name
      - email
      - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: Admin
        email: admin@gmail.com
        password: '123456'
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...